override some stuff deep in registration to fix PMT #97589
The way it used to work, the code monkey-patched the django.auth.contrib.models.User
object's email attribute to shim in the encrypted email.

Unsurprisingly, monkey-patching the model broke in Django 1.7. Now, instead, we
are subclassing a few bits of django-registration to get it to use the
encrypted_email field on the profile for sending email to instead.

The advantage of this approach is that there is no monkey patching, which
makes it quite a bit cleaner.

There are two downsides. First, django-registration isn't as modular as it could
be, so we had to pull quite a bit of code in to override one small thing. We ought
to refactor to make this better and push it upstream.

Second, there may still be places lingering in other apps that expect to just
be able to send email to the address directly on the User object. Not monkey-patching
it means that we might stumble upon those later and need to subclass them
in a similar manner.