:white_check_mark: smoketest for lessc
recently, after moving nynjaetc to patches, we discovered that
styling was messed up (PMT #98865).

Problem was that patches (being trusty) wasn't getting
lessc installed via Salt (there were explicit sections
for precise and lucid since npm stuff is so finicky, but
we hadn't included trusty yet). That was quickly fixed.

This adds a smoketest to make sure the lessc executable
exists and is in our path. So in the future, if we deploy
to a new host where lessc isn't there, it will fail. Or,
if an update on a server breaks lessc or moves it somewhere
not in our path, we will know about it quickly.