Fix: Reset password email not being sent
Nynjaetc overrides the PasswordResetForm.save to account for encrypted emails, then calls the original save.
This approach broke with a Django upgrade when a "users_cache" member variable was removed from the base class.
Updating the override to account for the change. (monkey patching keeps on giving.)

The Django update that removes the users_cache was adopted in 1.6:
https://github.com/django/django/commit/2f4a4703e1931fadf5ed81387b26cf84caf5bef9#diff-e840e362abe9e625eee52d91897400bd